backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["caien8 <caien811_P@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
icecream = "^2.1.4"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"
pillow = "^11.1.0"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
